Add ReadWriteClient for automatic read/write routing
Add ReadWriteClient wrapper that automatically routes read operations to a replica and write operations to primary.

Reads to the replica fallback to the writer if the read fails.
